api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy

# Updated TensorFlow versions (compatible with Python 3.10+)
tensorflow==2.20.0
tensorflow-serving-api==2.19.1

